move non mainline scenarios to helper functions
